Added resend activation email

+++ activate.php
@@ -1,29 +1,86 @@
+<!DOCTYPE html>
+<html lang="en">
 <?php
 require('includes/config.php');
+// require($_SERVER['DOCUMENT_ROOT'].'/classes/phpmailer/Mail.php');
+$title = 'CU HvZ | ';
+?>
+<head>
+	<?php require('layout/header.php'); ?>
+</head>
+<body>
+<?php
+include 'layout/navbar.php';
+
+$errors = [];
 
 //collect values from the url
-$user_id = trim($_GET['x']);
-$active_token = trim($_GET['y']);
-
-//if id is number and the active token is not empty carry on
-if(is_numeric($user_id) && !empty($active_token)){
-	$result = activateAccount($user_id);
+$userID = trim($_GET['x']);
+$activeToken = trim($_GET['y']);
+$database = new Database();
+validateToken($database, $userID, $activeToken, $errors);
+if(empty($errors)){
+	$result = activateAccount($database, $userID, $errors);
 	if($result){
-
 		//redirect to login page
-		header('Location: login.php?action=active');
+		header('Location: profile.php?action=activated');
 		exit;
-
-	} else {
-		echo "Your account could not be activated. Please contact the mod team.";
 	}
-
 }
 
-function activateAccount($userID){
-	$database = new Database();
+
+function validateToken($database, $userID, $token, &$errors){
+	$query = "select * from tokens where token='$token'";
+	$result = $database->executeQueryFetch($query);
+	// token does not exist in database
+	if(!$result){
+		array_push($errors, "Activation token is not valid.");
+	}
+	// an error occured executing database query
+	else	if(isset($result["error"])){
+		array_push($errors, "Error occured validating token.");
+	}	else {
+		$tokenID = $result["id"];
+		// test if token has expired
+		$expired = $database->isTokenExpired($tokenID);
+		if($expired){
+			array_push($errors, "Token has expired.");
+		}else{
+			$userID = $result["user_id"];
+		}
+	}
+}
+
+function activateAccount($database, $userID, &$errors){
 	$result = $database->update("user_stats", "activated = true", "id = $userID");
+	if(isset($result["error"])){
+		array_push($errors, "Your account could not be activated. Please contact the mod team.");
+		return false;
+	}
 	$database->delete("tokens", "user_id = $userID AND token_type = 'ACTIVATION'");
-	return $result;
+	return true;
 }
+?>
+
+<div class="lightslide">
+
+<div class="container">
+
+	<div class="row">
+	    	<?php
+				if(isset($errors)){
+					foreach($errors as $error){
+						echo '<p class="bg-danger">'.$error.'</p>';
+					}
+				}
+				?>
+		</div>
+	</div>
+</div>
+
+</div>
+
+<?php
+//include header template
+require('layout/footer.php');
 ?>
